esp32-cam-face-detection/computer/face_detection/requirements.txt
numpy==1.23.4
opencv-python==4.6.0
pyserial==3.5
matplotlib==3.6.0
